Departments enity and its CRUD added

+++ src/main/java/org/gklsan/springboot/myjavalearning/repository/StudentRepository.java
@@ -6,4 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 // custom routes names
 //@RepositoryRestResource(path = "people")
 
-public interface StudentRepository extends JpaRepository<Student, Integer> {}
+public interface StudentRepository extends JpaRepository<Student, Long> {}
